device_storage/test_stat.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="device_storage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

// It is a requirement for an sdcard to be loaded, before this test will pass.
// This is accomplished by adding a --sdcard 20/50...MB to the command line

devicestorage_setup();
SpecialPowers.addPermission("device-storage", true, document);
SpecialPowers.addPermission("devicestorage", true, document);


function statSuccess(e) {
  ok(e.target.result.freeBytes  > 0, "free bytes should exist and be greater than zero");
  ok(e.target.result.totalBytes > 0, "total bytes should exist and be greater than zero");
  cleanup = storage.delete("a/b.png");
  cleanup.onsuccess = function() {
    devicestorage_cleanup();
  };
  cleanup.onerror = function()  {
    devicestorage_cleanup();
  }
}

function statError(e) {
  ok(false, "statError was called");
  devicestorage_cleanup();
}

var storage = navigator.getDeviceStorage("pictures");
ok(navigator.getDeviceStorage, "Should have getDeviceStorage");

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  devicestorage_cleanup();
}

function addSuccess(e) {
  request = storage.stat();
  ok(request, "Should have a non-null request");

  request.onsuccess = statSuccess;
  request.onerror = statError;
}

request = storage.addNamed(createRandomBlob('image/png'), "a/b.png");
request.onsuccess = addSuccess;
request.onerror = addError;

</script>
</pre>
</body>
</html>
